Incluye la funcionaidad de productosIncluye Crud de cientes y encripcion de contraseña.

+++ EcommerceAPI.Comunes/Classes/Contracts/Ecommerce/ClienteContract.cs
@@ -2,11 +2,13 @@
 {
     public class ClienteContract
     {
-        public int id_cliente { get; set; }
+        // No es necesidad que tenga el mismos nombres de campos de la tabla en este punto
+        // Esto es como se le presentara al usuario final.
+        public Int64 id_cliente { get; set; }
         public string nombre { get; set; }
         public string contrasena { get; set; }
         public string correo { get; set; }
         public string direccioncliente { get; set; }
-        public int telefono { get; set; }
+        public decimal telefono { get; set; }
     }
 }
